Simplify shorteners API to only include?
GitHub: https://github.com/altmetric/embiggen/issues/2

Allow users to supply their own object for
Embiggen::Configuration.shorteners that only responds to include? to
determine whether a URI is shortened or not.

This replaces the current Set implementation with a ShortenerList class
that has the same API (to keep testing simple) but exposes a custom
include? method.