Add 531 shorteners & remove misleading example
As raised in #28, the `ExpandEverything` example in the README is
misleading as it will inevitably raise a `BadShortenedURI` exception as
it attempts to follow _every_ URL including those that return a 200.

As Embiggen is designed specifically for high throughput cases where
following every URL is not feasible (e.g. processing thousands of links
in a short space of time), remove this example.