Don't read response bodies when making GET request
A downside of switching from HEAD requests to GET requests is that we
will automatically begin reading response bodies into memory when trying
to follow shortened links.

For redirects (that is, HTTP 301 or 302 responses), this isn't a big
problem because response bodies tend to be empty or very short but if a
bad link is followed which results in a huge response body (e.g. a disk
image or video), the library will be blocked as it reads the body off
the socket.

Avoid this by using our own Net::HTTP request class which uses the GET
method but does not read the body from the response. This will cause
Net::HTTP to make a GET request but close the connection before trying
to read the entire body into memory.